+++ src/inputs/kafka_input.rs
@@ -39,13 +39,18 @@ use super::{distribute, InputConfig, InputReader};
 ///   starting_offset (str): Can be "beginning" or "end". Delegates
 ///       where to resume if auto_commit is not enabled. Defaults to
 ///       "beginning".
+/// 
+/// Any additional configuration properties can be passed as kwargs. Note
+/// that consumer group settings will be ignored.
+/// See https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md
+/// for more options.
 ///
 /// Returns:
 ///
 ///   Config object. Pass this as the `input_config` argument to the
 ///   `bytewax.dataflow.Dataflow.input`.
 #[pyclass(module = "bytewax.inputs", extends = InputConfig)]
-#[pyo3(text_signature = "(brokers, topic, tail, starting_offset)")]
+#[pyo3(text_signature = "(brokers, topic, tail, starting_offset, additional_configs='**')")]
 pub(crate) struct KafkaInputConfig {
     #[pyo3(get)]
     pub(crate) brokers: Vec<String>,
@@ -55,6 +60,8 @@ pub(crate) struct KafkaInputConfig {
     pub(crate) tail: bool,
     #[pyo3(get)]
     pub(crate) starting_offset: String,
+    #[pyo3(get)]
+    pub(crate) additional_configs: Option<HashMap<String, String>>,
 }
 
 #[pymethods]
@@ -64,13 +71,15 @@ impl KafkaInputConfig {
         brokers,
         topic,
         tail = true,
-        starting_offset = "\"beginning\".to_string()"
+        starting_offset = "\"beginning\".to_string()",
+        additional_configs = "**"
     )]
     fn new(
         brokers: Vec<String>,
         topic: String,
         tail: bool,
         starting_offset: String,
+        additional_configs: Option<HashMap<String, String>>,
     ) -> (Self, InputConfig) {
         (
             Self {
@@ -78,18 +87,20 @@ impl KafkaInputConfig {
                 topic,
                 tail,
                 starting_offset,
+                additional_configs
             },
             InputConfig {},
         )
     }
 
-    fn __getstate__(&self) -> (&str, Vec<String>, String, bool, String) {
+    fn __getstate__(&self) -> (&str, Vec<String>, String, bool, String, Option<HashMap<String, String>>) {
         (
             "KafkaInputConfig",
             self.brokers.clone(),
             self.topic.clone(),
             self.tail,
             self.starting_offset.clone(),
+            self.additional_configs.clone(),
         )
     }
 
@@ -101,11 +112,12 @@ impl KafkaInputConfig {
 
     /// Unpickle from tuple of arguments.
     fn __setstate__(&mut self, state: &PyAny) -> PyResult<()> {
-        if let Ok(("KafkaInputConfig", brokers, topic, tail, starting_offset)) = state.extract() {
+        if let Ok(("KafkaInputConfig", brokers, topic, tail, starting_offset, additional_configs)) = state.extract() {
             self.brokers = brokers;
             self.topic = topic;
             self.tail = tail;
             self.starting_offset = starting_offset;
+            self.additional_configs = additional_configs;
             Ok(())
         } else {
             Err(PyValueError::new_err(format!(
@@ -179,6 +191,7 @@ impl KafkaInput {
         topic: &str,
         tail: bool,
         starting_offset: Offset,
+        additional_configs: &Option<HashMap<String, String>>,
         worker_index: usize,
         worker_count: usize,
         resume_state_bytes: Option<StateBytes>,
@@ -187,13 +200,23 @@ impl KafkaInput {
             .map(|resume_state_bytes| resume_state_bytes.de())
             .unwrap_or_default();
 
-        let eof = !tail;
-
-        let consumer: BaseConsumer = ClientConfig::new()
+        let eof = !tail; 
+        
+        let mut config = ClientConfig::new();
+        
+        config
             .set("group.id", "BYTEWAX_IGNORED")
             .set("enable.auto.commit", "false")
             .set("bootstrap.servers", brokers.join(","))
-            .set("enable.partition.eof", format!("{eof}"))
+            .set("enable.partition.eof", format!("{eof}"));
+
+        if let Some(args) = additional_configs {
+            for (key, value) in args.iter() {
+                config.set(key, value);
+            }
+        }
+
+        let consumer: BaseConsumer = config
             .create()
             .expect("Error building input Kafka consumer");
 
@@ -201,6 +224,7 @@ impl KafkaInput {
         if partition_count == 0 {
             panic!("Topic {} does not exist, please create first", topic);
         }
+
         let mut partitions = TopicPartitionList::new();
         for partition in distribute(0..partition_count, worker_index, worker_count) {
             let partition = KafkaPartition(partition);